+++ src/main/java/com/deux/duohaeduo/dto/riot/matchInfo/Style.java
@@ -1,6 +1,6 @@
-package com.deux.duohaeduo.dto.matchInfo;
+package com.deux.duohaeduo.dto.riot.matchInfo;
 
-import com.deux.duohaeduo.dto.matchInfo.enums.Description;
+import com.deux.duohaeduo.dto.riot.matchInfo.enums.Description;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.util.List;
