+++ src/main/java/com/deux/duohaeduo/dto/riot/matchInfo/enums/Description.java
@@ -1,4 +1,4 @@
-package com.deux.duohaeduo.dto.matchInfo.enums;
+package com.deux.duohaeduo.dto.riot.matchInfo.enums;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
