+++ src/main/java/com/deux/duohaeduo/dto/riot/matchInfo/enums/IndividualPosition.java
@@ -1,17 +1,19 @@
-package com.deux.duohaeduo.dto.matchInfo.enums;
+package com.deux.duohaeduo.dto.riot.matchInfo.enums;
+
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
 
 import java.io.IOException;
 
-public enum Lane {
+public enum IndividualPosition {
 
-    BOTTOM, MIDDLE, TOP, JUNGLE, NONE, UTILITY;
+    INVALID, NONE, JUNGLE, TOP, MIDDLE, BOTTOM, UTILITY;
 
     @JsonValue
     public String toValue() {
         return switch (this) {
+            case INVALID -> "Invalid";
             case BOTTOM -> "BOTTOM";
             case MIDDLE -> "MIDDLE";
             case TOP -> "TOP";
@@ -22,14 +24,15 @@ public enum Lane {
     }
 
     @JsonCreator
-    public static Lane forValue(String value) throws IOException {
+    public static IndividualPosition forValue(String value) throws IOException {
         if (value.equals("BOTTOM")) return BOTTOM;
         if (value.equals("MIDDLE")) return MIDDLE;
         if (value.equals("TOP")) return TOP;
         if (value.equals("JUNGLE")) return JUNGLE;
         if (value.equals("UTILITY")) return UTILITY;
         if (value.equals("NONE")) return NONE;
-        throw new IOException("Cannot deserialize Lane");
+        if (value.equals("Invalid")) return INVALID;
+        throw new IOException("Cannot deserialize IndividualPosition");
     }
 
 }
